# apps/katsubot-web/public/api-docs/specs/openapi.yaml
openapi: 3.1.0
info:
  title: Katsubot API
  version: 0.2.0
  description: |
    katsubot-api BFF — browser ↔ katsubot-api 계약 (Phase 2).
    RAG 호출은 katsubot-api 내부 Port이며 본 스펙에 포함되지 않음.
    Phase 3 운영 RAG·OTel: `docs/modernization/rag-ops-contract.md`, `docs/rag-external-client.md`.
servers:
  - url: http://localhost:8081
    description: 로컬 katsubot-api
  - url: http://localhost:5173/api
    description: Vite dev proxy (`/api` → katsubot-api)
security:
  - bearerAuth: []
paths:
  /api/v1/conversations:
    post:
      operationId: createConversation
      summary: 새 대화 생성
      tags: [conversations]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationRequest'
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      operationId: listConversations
      summary: 내 대화 목록
      tags: [conversations]
      responses:
        '200':
          description: 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversation'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      operationId: deleteConversations
      summary: 대화 일괄 삭제
      description: |
        사용자는 JWT `sub`로 식별한다. 본인 소유 대화만 삭제 가능.
        레거시 `DELETE /xs/aichat/v2/conversations` 대응.
      tags: [conversations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteConversationsRequest'
      responses:
        '200':
          description: 삭제 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteConversationsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/conversations/{conversationId}/messages:
    get:
      operationId: listMessages
      summary: 대화 메시지 히스토리
      description: 레거시 `GET /xs/aichat/v2/messages` 대응. 커서 기반 페이지네이션.
      tags: [messages]
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 메시지 페이지
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: sendMessage
      summary: 메시지 전송 (SSE 스트림)
      tags: [messages]
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: |
            SSE 스트림. 각 이벤트는 `data: {json}` 형식.
            - `event: delta` — `{"delta":"토큰"}`
            - `event: done` — `{"conversation_id":"...","message_id":"...","title":"..."}` (`title` optional, 첫 메시지 후 Gateway LLM 제목)
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  event: delta
                  data: {"delta":"안녕"}

                  event: done
                  data: {"conversation_id":"c1","message_id":"m1","title":"휴가 규정 문의"}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/conversations/{conversationId}/messages/{messageId}/feedback:
    put:
      operationId: upsertMessageFeedback
      summary: 메시지 피드백 저장·수정
      description: 레거시 `PUT .../feedback` 대응. `like` / `dislike`.
      tags: [feedback]
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/MessageId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertFeedbackRequest'
      responses:
        '200':
          description: 저장됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feedback'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/conversations/{conversationId}/messages/{messageId}/feedback/{feedbackId}:
    delete:
      operationId: deleteMessageFeedback
      summary: 메시지 피드백 취소
      description: 레거시 `DELETE .../feedback/{feedback_id}` 대응 (논리 삭제).
      tags: [feedback]
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/MessageId'
        - $ref: '#/components/parameters/FeedbackId'
      responses:
        '200':
          description: 취소됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feedback'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/auth/encrypt-key:
    post:
      operationId: createLoginEncryptKey
      summary: 로그인 OTP 암호화 키 발급
      description: |
        레거시 `POST /xs/vob/aichat/auth/encrypt-key` 대응.
        세션에 일회용 AES 키를 저장하고 클라이언트에 전달한다.
      tags: [auth]
      security: []
      responses:
        '200':
          description: 암호화 키
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncryptKeyResponse'
        '503':
          description: hyobee-admin-db 미설정
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      operationId: loginWithPassword
      summary: ID/비밀번호 로그인
      description: |
        레거시 `POST /xs/vob/aichat/auth/login` 대응.
        hyobee-admin-db(`com_user` 등)에서 사용자를 검증하고 Hyobee 호환 JWT를 발급한다.
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: JWT 발급
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 로그인 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: hyobee-admin-db 미설정
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/board-auth:
    get:
      operationId: listBoardAuth
      summary: 접근 가능 게시판 목록
      description: |
        레거시 `GET /xs/aichat/v2/board-auth` 대응.
        Phase 2는 `BoardAuthPort` + 레거시 브릿지 (`docs/modernization/board-auth-bridge.md`).
      tags: [board-auth]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 게시판 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardAuthPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 레거시 SSO/JWT 브릿지 — `docs/auth-bridge.md`
  parameters:
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MessageId:
      name: messageId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FeedbackId:
      name: feedbackId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: |
        메시지 목록 커서. 미지정 시 첫 페이지.
        in-memory/jpa — 숫자 오프셋 문자열(예: "20").
        gateway 프로필 — Gateway `next_cursor`(message UUID).
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    PageSize:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    CreateConversationRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          example: 새 대화
        chat_category:
          type: string
          description: internal_rules, web_search, rnd_search
          example: internal_rules
    Conversation:
      type: object
      required: [id, title, created_at]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        created_at:
          type: string
          format: date-time
        chat_category:
          type: string
          description: internal_rules, web_search, rnd_search 등 (레거시 대화 목록 뱃지용)
          example: internal_rules
    DeleteConversationsRequest:
      type: object
      required: [conversation_ids]
      properties:
        conversation_ids:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
    DeleteConversationsResponse:
      type: object
      required: [deleted_count, results]
      properties:
        deleted_count:
          type: integer
          minimum: 0
        results:
          type: array
          items:
            $ref: '#/components/schemas/DeleteConversationResult'
    DeleteConversationResult:
      type: object
      required: [conversation_id, deleted]
      properties:
        conversation_id:
          type: string
          format: uuid
        deleted:
          type: boolean
        error:
          type: string
    SendMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 8000
          example: 안녕하세요
        chat_category:
          type: string
          description: internal_rules, web_search, rnd_search
          example: internal_rules
    Message:
      type: object
      required: [id, role, content, created_at]
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        created_at:
          type: string
          format: date-time
        feedback:
          $ref: '#/components/schemas/MessageFeedbackSummary'
    MessageFeedbackSummary:
      type: object
      required: [feedback_id, feedback_type]
      properties:
        feedback_id:
          type: string
          format: uuid
        feedback_type:
          $ref: '#/components/schemas/FeedbackType'
    MessagesPage:
      type: object
      required: [messages, has_more, next_cursor]
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        has_more:
          type: boolean
        next_cursor:
          type: string
          nullable: true
          description: 다음 페이지 커서 (오프셋 문자열 또는 message UUID)
    UpsertFeedbackRequest:
      type: object
      required: [feedback_type]
      properties:
        feedback_type:
          $ref: '#/components/schemas/FeedbackType'
    FeedbackType:
      type: string
      enum: [like, dislike]
    Feedback:
      type: object
      required: [feedback_id, message_id, feedback_type]
      properties:
        feedback_id:
          type: string
          format: uuid
        message_id:
          type: string
          format: uuid
        feedback_type:
          $ref: '#/components/schemas/FeedbackType'
    BoardAuthPage:
      type: object
      required: [items, total_elements, page_number, page_size, has_next]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BoardAuthItem'
        total_elements:
          type: integer
          minimum: 0
        page_number:
          type: integer
          minimum: 0
        page_size:
          type: integer
          minimum: 0
        has_next:
          type: boolean
    BoardAuthItem:
      type: object
      required: [board_name]
      properties:
        board_name:
          type: string
    EncryptKeyResponse:
      type: object
      required: [encrypt_key]
      properties:
        encrypt_key:
          type: string
          description: 32자 AES OTP 키 (Base64 prefix)
    LoginRequest:
      type: object
      required: [company_code_encrypt, user_id_encrypt, password_encrypt]
      properties:
        company_code_encrypt:
          type: string
        user_id_encrypt:
          type: string
        password_encrypt:
          type: string
        language_code:
          type: string
          default: ko
    LoginResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: Hyobee 호환 HS512 JWT
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
  responses:
    Unauthorized:
      description: 인증 필요 또는 토큰 무효
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Authorization Bearer 토큰이 필요합니다
    Forbidden:
      description: 리소스 소유자 불일치
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: 대화에 접근할 권한이 없습니다
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: 대화를 찾을 수 없습니다
    BadRequest:
      description: 요청 검증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: BAD_REQUEST
            message: 요청이 유효하지 않습니다